added architecture diagram
