+++ common/README.md
@@ -1,0 +1,9 @@
+# About
+ 
+Library of common components used by all other modules. It:
+
+* Pulls in JARS for the Apache Kafka client library, Logback, JUnit 5 etc.
+* Provides a config file and helper class for accessing properties defined in a Maven profile.
+* Contains the Logback configuration file.
+* Defines the SerDes classes used by Kafka clients.
+* Defines common DTOs used to pass data between components of the pipeline.
