+++ dashboard/pom.xml
@@ -5,18 +5,17 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.2.6.RELEASE</version>
+		<version>2.3.1.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>co.adeshina.c19ta</groupId>
-	<artifactId>data-api</artifactId>
+	<artifactId>dashboard</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
-	<name>Data API</name>
+	<name>Dashboard</name>
+	<description>Frontend of the project</description>
 
 	<properties>
 		<java.version>1.8</java.version>
-		<commom.version>1.0.0-SNAPSHOT</commom.version>
-		<kafka.clients.version>2.5.0</kafka.clients.version>
 	</properties>
 
 	<dependencies>
@@ -27,30 +26,6 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-test</artifactId>
-			<scope>test</scope>
-			<exclusions>
-				<exclusion>
-					<groupId>org.junit.vintage</groupId>
-					<artifactId>junit-vintage-engine</artifactId>
-				</exclusion>
-			</exclusions>
-		</dependency>
-
-		<dependency>
-			<groupId>co.adeshina.c19ta</groupId>
-			<artifactId>common</artifactId>
-			<version>${commom.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.apache.kafka</groupId>
-			<artifactId>kafka-clients</artifactId>
-			<version>${kafka.clients.version}</version>
 		</dependency>
 
 	</dependencies>
