+++ dashboard/src/main/java/co/adeshina/c19ta/dashboard/UserInterfaceApplication.java
@@ -1,0 +1,13 @@
+package co.adeshina.c19ta.dashboard;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserInterfaceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserInterfaceApplication.class, args);
+	}
+
+}
