+++ dashboard/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+server.port=@dashboard.port@
+data.api.url=@data.api.url@
