+++ data-api/README.md
@@ -1,0 +1,5 @@
+# About
+ 
+A Spring Boot web application that polls a Kafka broker for records written by the stream processing component of the
+pipeline. It then transforms those records into data packages that are pushed to the dashboard in server-sent
+events.
