+++ data-api/src/main/java/co/adeshina/c19ta/data_api/DataApiApplication.java
@@ -1,0 +1,13 @@
+package co.adeshina.c19ta.data_api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DataApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DataApiApplication.class, args);
+	}
+
+}
