--- data-api/src/main/java/co/adeshina/c19ta/datavisualizer/DataApiApplication.java
@@ -1,13 +1,0 @@
-package co.adeshina.c19ta.datavisualizer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DataApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DataApiApplication.class, args);
-	}
-
-}
