+++ data-api/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=@api.server.port@
+server.port=@data.api.port@
 
 # Id of the group the consumer is a member of.
 group.id=data-api
