+++ extractor/README.md
@@ -1,0 +1,9 @@
+# About
+ 
+ This a Java program that opens a persistent HTTP connection to Twitter's 
+ [Filtered Stream API*](https://developer.twitter.com/en/docs/labs/filtered-stream/overview) from which it consumes
+ tweets as they are sent in realtime and writes tweet metadata to a Kafka broker.
+ 
+ 
+ *Only available in the Twitter Developer Labs program as of August 2020.
+
