--- user-interface/src/main/java/co/adeshina/c19ta/ui/UserInterfaceApplication.java
@@ -1,13 +1,0 @@
-package co.adeshina.c19ta.ui;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserInterfaceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserInterfaceApplication.class, args);
-	}
-
-}
