--- user-interface/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-server.port=@user.interface.port@
-data.api.url=@data.api.url@
