--- user-interface/src/main/resources/templates/dashboard.html
@@ -1,23 +1,0 @@
-<!Doctype html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>COVID-19 Dashboard</title>
-    <link rel="stylesheet" th:href="@{/css/dashboard.css}">
-    <script th:src="@{https://cdn.jsdelivr.net/npm/chart.js@2.9.3/dist/Chart.min.js}"></script>
-</head>
-
-<body>
-    
-    <input type="hidden" id="api_url" th:value="${dataApiUrl}">
-
-    <div id="container">
-        <canvas id="chart"></canvas>
-    </div>
-
-    <!-- todo: Display total number of tweets here -->
-
-    <script th:src="@{/js/dashboard.js}"></script>
-</body>
-</html>
